Organized scripts better
To give the template a more modular structiure, objects that do not need to be in MAIN.SC have been moved into their own scripts.

Also, sci2.sh has been converted to Sierra Script, but the defines should be considered deprecated, and have been kept for compatibility purposes. It is now recommended to use the defines in SYSTEM.SH and KERNEL.SH.